+++ jobs/fetch-kaggle/vitest.config.ts
@@ -14,11 +14,11 @@ export default defineConfig({
       reportsDirectory: '../../coverage/jobs/fetch-kaggle',
       provider: 'v8',
       reporter: ['text', 'lcov', 'html'],
-      include: ['src/lib/utils/**/*.ts'],
+      include: ['src/lib/utils/**/*.ts', 'src/lib/auth/**/*.ts'],
       thresholds: {
         lines: 90,
         functions: 90,
-        branches: 90,
+        branches: 85,
         statements: 90,
       },
     },